backend_p12/requirements.txt
# Project Tinder Backend - Requirements for Pages 1 & 2
# Install: pip install -r requirements.txt
# Deploy: scp -r ~/Desktop/backend_p12 ubuntu@YOUR_CVM_IP:/home/ubuntu/

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation  
pydantic>=2.9.0

# HTTP requests for external APIs
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# PostgreSQL database - psycopg3 for Python 3.13 compatibility
psycopg[binary]==3.1.8
sqlalchemy>=2.0.25

# Database migrations
alembic==1.13.0

# JWT and Authentication
python-jose[cryptography]==3.3.0

# Password hashing
passlib[bcrypt]==1.7.4

# Email validation for Pydantic EmailStr
email-validator==2.0.0

# Rate limiting (optional but recommended)
slowapi==0.1.9

# Vector embeddings (optional - for local query embedding)
sentence-transformers>=2.7.0
huggingface_hub>=0.19.0

# Tencent Cloud Vector Database (optional)
tcvectordb